:triangular_ruler: Script subsystem tidy-up
Renamed ScriptEngine to script-engine, for consistency.
Extracted 'server-script' from 'script-engine', moved class ServerScript
added function `static void ServerScript::Register()`, code extracted from `ScriptEngine::Init()`
